pyproject.toml
[tool.poetry]
name = "funny-dolphine"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.1.0"
alembic = "^1.12.1"
fastapi = "^0.104.1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
uvicorn = "^0.25.0"
itsdangerous = "^2.1.2"
sqladmin = "^0.16.0"
bcrypt = "^4.1.2"
pydantic = {extras = ["email"], version = "^2.5.3"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"
black = "^23.12.1"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
